requirements.txt
streamlit
supabase
pandas
altair
pillow
pytesseract
